ka-10840-add-metrics: send packageType in UserAgent
