+++ src/Kontur.Extern.Api.Client.Http/Kontur.Extern.Api.Client.Http.csproj
@@ -7,6 +7,13 @@
         <Configurations>Debug;Release;Client</Configurations>
     </PropertyGroup>
 
+    <PropertyGroup>
+        <IsNugetPackage Condition="'$(IsNugetPackage)' == ''">false</IsNugetPackage>
+    </PropertyGroup>
+    <ItemGroup>
+        <AssemblyMetadata Include="IsNugetPackage" Value="$(IsNugetPackage)" />
+    </ItemGroup>
+    
     <ItemGroup>
       <PackageReference Include="JetBrains.Annotations" Version="2021.2.0" />
       <PackageReference Include="Microsoft.AspNetCore.JsonPatch" Version="6.0.10" />
