--- src/Kontur.Extern.Api.Client.Http/VersionGetter/ClientVersionGetter.cs
@@ -1,19 +1,0 @@
-using System;
-using System.Reflection;
-
-namespace Kontur.Extern.Api.Client.Http.VersionGetter;
-
-public static class ClientVersionGetter
-{
-    public static readonly Lazy<string> ClientVersion = new(GetClientVersion);
-    
-    private static string GetClientVersion()
-    {
-        var version = Assembly
-            .GetExecutingAssembly()
-            .GetCustomAttribute<AssemblyInformationalVersionAttribute>()
-            .InformationalVersion;
-        
-        return version;
-    }
-}
